Github login system and delete working

+++ public/media/js/nc-subdomainer.js
@@ -1,18 +1,25 @@
+var inFlight = false;
 $(document).ready(function(){
     reloadHosts();
 
     // Add a new subdomain
     $("#addButton").click(function(evt){
-        $.ajax({
-            dataType: "json",
-            type: "post",
-            url: "/subdomain",
-            data: $("#addForm").serialize(),
-            success: function(res){
-                alert(res.message);
-                reloadHosts();
-            }
-        });
+        if (!inFlight){
+            inFlight = true;
+            $("#loadingRow").show();
+            $.ajax({
+                dataType: "json",
+                type: "post",
+                url: "/subdomain",
+                data: $("#addForm").serialize(),
+                success: function(res){
+                    inFlight = false;
+                    $("#loadingRow").hide();
+                    alert(res.message);
+                    reloadHosts();
+                }
+            });
+        }
     });
 
     // Change placeholder on dropdown update
@@ -66,23 +73,32 @@ var reloadHosts = function(){
 
                         rowOutput = rowOutput + "<td>" + records.host[a].Address + "</td>";
                         rowOutput = rowOutput + '<td><button type="button" class="btn btn-danger deleteButton" data-domain="' + records.Domain + '" data-subdomain="' + records.host[a].Name + '">Delete</button></td>';
-                        $("#HolderTable").append(rowOutput);
+                        $("#HolderTable tbody").append(rowOutput);
                     }
                 }
             }
-
+            
+            $("#HolderTable").tablesorter();
+            
             // Delete a new subdomain
             $(".deleteButton").click(function(evt){
-                var data = "subdomain=" + $(this).data("subdomain") + "&domain=" + $(this).data("domain");
-                $.ajax({
-                    dataType: "json",
-                    type: "delete",
-                    url: "/subdomain",
-                    "data": data,
-                    success: function(res){
-                        console.log(res);
-                    }
-                });
+                if (!inFlight){
+                    $("#loadingRow").show();
+                    inFlight = true;
+                    var data = "subdomain=" + $(this).data("subdomain") + "&domain=" + $(this).data("domain");
+                    $.ajax({
+                        dataType: "json",
+                        type: "delete",
+                        url: "/subdomain",
+                        "data": data,
+                        success: function(res){
+                            $("#loadingRow").hide();
+                            inFlight = false;
+                            console.log(res);
+                            reloadHosts();
+                        }
+                    });
+                }
             });
         }
     });
